ownfeauture

+++ eFlow.Application/Services/FlowerService.cs
@@ -1,23 +1,45 @@
 ﻿
+using eFlow.Core.Interfaces.Files;
 using eFlow.Core.Models;
 using eFlow.Infrastructure.Repositories;
+using Microsoft.AspNetCore.Http;
 
 namespace eFlow.Application.Services
 {
     public class FlowerService
     {
         private readonly IFlowerRepository _flowerRepository;
-        public FlowerService(IFlowerRepository flowerRepository)
+        private readonly IMediaFileService _mediaFileService;
+        public FlowerService(IFlowerRepository flowerRepository, IMediaFileService mediaFileService)
         {
             _flowerRepository = flowerRepository;
+            _mediaFileService = mediaFileService;
         }
-        public async Task<Guid> CreateFlowerAsync(Flower flower)
+        public async Task<Guid> CreateFlowerAsync(Flower flower, IFormFile? imageFile)
         {
+            if (imageFile != null)
+            {
+                var mediaFile = await _mediaFileService.UploadAsync(imageFile);
+                flower.ImageFile = mediaFile;
+            }
             var data = await _flowerRepository.AddAsync(flower);
             return data;
         }
-        public async Task<bool> UpdateFlowerAsync(Flower flower)
+        public async Task<bool> UpdateFlowerAsync(Flower flower, IFormFile? imageFile)
         {
+            var existFlower = await _flowerRepository.GetByIdAsync(flower.Id);
+            if(existFlower == null) return false;
+            
+            if (imageFile != null)
+            {
+                if (existFlower.ImageFile != null)
+                {
+                    await _mediaFileService.DeleteAsync(existFlower.ImageFile.FileName);
+                }
+                var mediaFile = await _mediaFileService.UploadAsync(imageFile);
+                flower.ImageFile = mediaFile;
+            }
+
             var data = await _flowerRepository.UpdateAsync(flower);
             return data;
         }
@@ -38,8 +60,17 @@ namespace eFlow.Application.Services
         }
         public async Task<bool> DeleteAsync(Guid id)
         {
-            var data = await _flowerRepository.RemoveAsync(id);
-            return data;
+            var existFlower = await _flowerRepository.GetByIdAsync(id);
+            if (existFlower != null)
+            {
+                if (existFlower.ImageFile != null)
+                {
+                    await _mediaFileService.DeleteAsync(existFlower.ImageFile.FileName);
+                }
+                var data = await _flowerRepository.RemoveAsync(id);
+                return data;
+            }
+            return false;           
         }
     }
 }
